# resources/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: "Bazorg"
  version: "1.0.0"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/v1/register:
    post:
      operationId: "register a user"
      tags:
        - "User Auth"
      summary: "User register"
      description: "register by email and password Here"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - email
                - password
                - password_confirmation
              properties:
                email:
                  type: string
                  example: "m@gmail.com"
                password:
                  type: string
                  example: "123mM!"
                password_confirmation:
                  type: string
                  example: "123mM!"
      responses:
        '201':
          description: "register Successfully"
          content:
            application/json:
              schema:
                type: object
        '200':
          description: "register Successfully"
          content:
            application/json:
              schema:
                type: object
        '422':
          description: "Unprocessable Entity"
          content:
            application/json:
              schema:
                type: object
        '400':
          description: "Bad request"
        '404':
          description: "Resource Not Found"
        '500':
          description: "Server"
...